Common: Move StringPredicateFn from pipeline and fix imports
